refactor: constants 파일 기능별로 세분화 작업
