+++ app/(routes)/[genre]/page.tsx
@@ -1,8 +1,8 @@
+import { FIRST_PAGE } from "@/app/modules/config";
 import Error from "../../components/@common/Error";
 import GenreContents from "../../components/@common/GenreContents";
 import { MusicLayout } from "../../components/@common/MusicLayout";
 import { fetchGenreData } from "../../modules/api/album";
-import { FIRST_PAGE } from "../../modules/constants";
 import { PageProps } from "../../modules/types";
 
 export default async function Page({ params }: PageProps): Promise<React.ReactElement> {
