+++ app/(routes)/page.tsx
@@ -2,7 +2,7 @@ import Error from "../components/@common/Error";
 import { MusicLayout } from "../components/@common/MusicLayout";
 import { LandingPage } from "../components/landingPage/LandingPage";
 import { fetchAlbumDataSSR } from "../modules/api/album";
-import { PER_PAGE_COUNT } from "../modules/constants";
+import { PER_PAGE_COUNT } from "../modules/config";
 
 export default async function Page() {
   try {
