+++ app/modules/api/auth.ts
@@ -1,7 +1,7 @@
 import { toast } from "react-toastify";
-import { BASE_URL } from "../constants";
 import { validateEmail, validatePassword, validateUserId, validateUserName } from "../utils";
 import { verify } from "jsonwebtoken";
+import { BASE_URL } from "../constants/apiUrls";
 
 export async function userSignUp(
   userId: string,
