+++ app/modules/config/index.ts
@@ -1,0 +1,5 @@
+export const SITE_TITLE: string = "카버뮤직";
+export const FIRST_PAGE: number = 1;
+export const MIN_SCROLL_COUNT: number = 1;
+export const PER_PAGE_COUNT: number = 50;
+export const SUB_PER_PAGE_COUNT: number = 5;
