+++ app/modules/constants/apiUrls.ts
@@ -1,0 +1,3 @@
+require("dotenv").config();
+
+export const BASE_URL = process.env.NEXT_PUBLIC_BASE_URL;
