+++ app/modules/constants/genres.ts
@@ -1,0 +1,18 @@
+import { GenreTranslations } from "../types";
+
+export const GENRES: GenreTranslations = {
+  pop: "팝",
+  kpop: "케이팝",
+  jpop: "제이팝",
+  rock: "록",
+  alternative: "얼터너티브",
+  disco: "디스코",
+  electronic: "일렉트로닉",
+  jazz: "재즈",
+  soul: "알앤비/소울",
+  hiphop: "힙합",
+  folk: "포크",
+  country: "컨트리",
+  classic: "클래식",
+  soundtrack: "사운드트랙",
+};
